Improve interpreting benchmark JSON results
In case a benchmark errors out, do not include it in the benchmark
results and graphs.

Furthermore, do not weigh the overall relative performance according to
the number of models per format type; this would give an unfair
advantage to libraries that would have strong performance for some
format type for which more models are available than other format types.